feat: add app storybook config
